IdKeyer: clean code, remove message and set version to 1.0 issue #239

+++ plugins/image/process/color/IdKeyer/src/IdKeyerDefinitions.hpp
@@ -13,7 +13,7 @@ static const std::string kParamNbPoints = "nbPoints";
 static const size_t      kMaxNbPoints   = 5;
 static const std::string kParamColor    = "color";
 
-inline std::string getColorParamName( const unsigned int i )
+inline std::string getColorParamName( const size_t i )
 {
 	return kParamColor + boost::lexical_cast<std::string>( i );
 }
